pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "magefight"
dependencies = [
  "requests",
  "click",
  "django",
  "djangorestframework",
]
version = "1.0.0"

[project.scripts]
magefight = "magefight.cli:main"
